fixing linter errors in ui button
